check if power of 2
in this code i use bitwise to check if number is a power of 2. in electronics all systems and computers are built upon the power of 2. for example we have all heard of  16, 32, 64 bits. these are all power of 2's.
we use bitwise logic because its the quickest way and most efficient. 
in the given i gave a none power of 2 number.